# docs/swagger.yaml
definitions:
  models.APIResponse:
    properties:
      data:
        description: Optional data payload
      error:
        description: |-
          Optional error
          example: Something went wrong
        type: string
      message:
        description: |-
          The response message
          example: Success
        type: string
    type: object
  models.AssignedSoftware:
    properties:
      assigned_at:
        type: string
      id:
        example: 1
        type: integer
      software_id:
        example: 3
        type: integer
      source:
        example: department
        type: string
      staff_id:
        example: 2
        type: integer
      updated_at:
        type: string
    type: object
  models.Department:
    properties:
      created_at:
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Engineering
        type: string
      updated_at:
        type: string
    type: object
...
  models.Team:
    properties:
      created_at:
        type: string
      department:
        $ref: '#/definitions/models.Department'
      department_id:
        example: 2
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: Frontend
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/assign-software:
    post:
      consumes:
      - application/json
      parameters:
      - description: Software assignment payload
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/models.AssignedSoftware'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AssignedSoftware'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Assign software to a staff member
      tags:
      - Assigned Software
  /api/assigned-software:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AssignedSoftware'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get all assigned software records
      tags:
      - Assigned Software
  /api/assigned-software/{id}:
    delete:
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: No Content - Deletion Success
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Delete assigned software and log the unassignment
      tags:
      - Assigned Software
    get:
      parameters:
      - description: Assigned Software ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AssignedSoftware'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get a specific assigned software record by ID
      tags:
      - Assigned Software
    put:
      consumes:
      - application/json
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated assignment
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/models.AssignedSoftware'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AssignedSoftware'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Update an assigned software record
      tags:
      - Assigned Software
  /api/assigned-software/{id}/force:
    delete:
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content - Deletion Success
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Delete an assigned software record
      tags:
      - Assigned Software
  /api/departments:
    get:
      parameters:
      - description: Search by name
        in: query
        name: search
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Department'
            type: array
      summary: List all departments with filters
      tags:
      - Departments
    post:
      consumes:
      - application/json
      parameters:
      - description: Department to create
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/models.Department'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Department'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Create a new department
      tags:
      - Departments
  /api/departments/{id}:
    delete:
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: No Content - Deletion Success
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Delete a department by ID
      tags:
      - Departments
    put:
      consumes:
      - application/json
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated department
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/models.Department'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Department'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Update a department by ID
      tags:
      - Departments
  /api/departments/{id}/teams:
    get:
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Team'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: List teams under a specific department
      tags:
      - Departments
...
  /api/staff/{id}/offboard:
    post:
      parameters:
      - description: Staff ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Offboard staff (revoke all software, optionally mark as inactive)
      tags:
      - Staff
...
swagger: "2.0"